+++ app/src/main/java/com/imashnake/animite/dev/SharedContentKey.kt
@@ -1,0 +1,16 @@
+package com.imashnake.animite.dev
+
+data class SharedContentKey(
+    val id: Int? = null,
+    val source: String? = null,
+    val sharedComponents: Pair<Component, Component>
+) {
+    enum class Component {
+        Card,
+        Page,
+        Text,
+        Image,
+    }
+
+    override fun toString() = "${sharedComponents.first}_${sharedComponents.second}_${id}_${source}"
+}
