+++ app/src/main/java/com/imashnake/animite/features/MainActivity.kt
@@ -155,18 +155,3 @@ fun MainScreen(modifier: Modifier = Modifier) {
         }
     }
 }
-
-data class SharedContentKey(
-    val id: Int? = null,
-    val source: String? = null,
-    val sharedComponents: Pair<Component, Component>
-) {
-    enum class Component {
-        Card,
-        Page,
-        Text,
-        Image,
-    }
-
-    override fun toString() = "${sharedComponents.first}_${sharedComponents.second}_${id}_${source}"
-}
