+++ app/src/main/java/com/imashnake/animite/features/home/HomeScreen.kt
@@ -70,11 +70,11 @@ import com.imashnake.animite.core.ui.MediaSmallRow
 import com.imashnake.animite.core.ui.ProgressIndicator
 import com.imashnake.animite.core.ui.layouts.BannerLayout
 import com.imashnake.animite.core.ui.layouts.TranslucentStatusBarLayout
-import com.imashnake.animite.features.SharedContentKey
-import com.imashnake.animite.features.SharedContentKey.Component.Card
-import com.imashnake.animite.features.SharedContentKey.Component.Image
-import com.imashnake.animite.features.SharedContentKey.Component.Page
-import com.imashnake.animite.features.SharedContentKey.Component.Text
+import com.imashnake.animite.dev.SharedContentKey
+import com.imashnake.animite.dev.SharedContentKey.Component.Card
+import com.imashnake.animite.dev.SharedContentKey.Component.Image
+import com.imashnake.animite.dev.SharedContentKey.Component.Page
+import com.imashnake.animite.dev.SharedContentKey.Component.Text
 import com.imashnake.animite.features.media.MediaPage
 import kotlinx.serialization.Serializable
 import com.imashnake.animite.core.R as coreR
